Download and modify logo files if needed

+++ build.sh
@@ -19,8 +19,16 @@ done
 
 OUTPUT_BASE_NAME="data_management_guidance_in_the_field"
 
-curl "https://swisspolar.ch/wp-content/uploads/2019/10/Logo-Swiss-Polar-Institute-HD.png" -o spi_logo_hd.png
-curl "https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" -o by.png
+SPI_LOGO_HD_FILE=spi_logo_hd.png
+if [ ! -f "$SPI_LOGO_HD_FILE" ]
+then
+	curl "https://swisspolar.ch/wp-content/uploads/2019/10/Logo-Swiss-Polar-Institute-HD.png" | convert -density 300 -units pixelsperinch - "$SPI_LOGO_HD_FILE"
+fi
+
+if [ ! -f "by.png" ]
+then
+	curl "https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" -o by.png
+fi
 
 # Generate tex file output using a template
 pandoc --pdf-engine=xelatex \
